Add Test, clean up and create FakeGameInitializer class
