+++ src/IM.Production.UI/src/contracts/contracts.component.spec.ts
@@ -2,16 +2,41 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { DebugElement } from '@angular/core';
-
+import { HttpClient } from '@angular/common/http';
 import { ContractsComponent } from './contracts.component';
 
 describe('ContractsComponent', () => {
   let component: ContractsComponent;
   let fixture: ComponentFixture<ContractsComponent>;
+  let httpClientMock: HttpClient = {
+    /*
+    handler:null,
+    request:null,
+    delete:null,
+    get:null,
+    post:null,
+    head:null,jsonp:null,options:null,patch:null,put:null
+    */  
+    /*
+    handler: jasmine.createSpy('handler'),
+    request: jasmine.createSpy('request'),
+    delete: jasmine.createSpy('request'),
+    get: jasmine.createSpy('get'),
+    post: jasmine.createSpy('post'),
+    head: jasmine.createSpy('head'),
+    jsonp: jasmine.createSpy('jsonp'),
+    options: jasmine.createSpy('options'),
+    patch: jasmine.createSpy('options'),
+    put: jasmine.createSpy('options')
+    */
+  };
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ ContractsComponent ]
+      declarations: [ ContractsComponent ],
+      providers: [
+        { provide: HttpClient, useValue: httpClientMock }
+      ]
     })
     .compileComponents();
   }));
@@ -25,4 +50,9 @@ describe('ContractsComponent', () => {
   it('should create', () => {
     expect(component).toBeTruthy();
   });
+
+  //it('weird stuff', () => {
+  //  expect(component.arrayData.length).toBeCloseTo(1);
+  //});
+
 });
