+++ src/IM.Production.WebApp.Tests/ContractControllerTests.cs
@@ -1,0 +1,18 @@
+﻿using IM.Production.WebApp.Controllers;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace IM.Production.WebApp.Tests
+{
+    [TestClass]
+    public class ContractControllerTests
+    {
+        [TestMethod]
+        public void GetAllContracts()
+        {
+            var contractsController = new ContractsController();
+            var allContracts = contractsController.GetAllContracts();
+            Assert.IsNotNull(allContracts);
+        }
+
+    }
+}
